config module *.svg globally to use svg images
