+++ package.json
@@ -9,33 +9,35 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/metro-runtime": "~4.0.1",
+    "@expo/metro-runtime": "~5.0.4",
     "@react-native-community/datetimepicker": "^8.3.0",
     "@react-navigation/bottom-tabs": "^7.2.1",
-    "@react-navigation/native": "^7.0.15",
+    "@react-navigation/native": "^7.1.9",
     "@react-navigation/native-stack": "^7.3.11",
     "eventease-mobileteamsenpai": "file:",
-    "expo": "~52.0.38",
-    "expo-image-picker": "~16.0.6",
+    "expo": "^53.0.0",
+    "expo-image-picker": "~16.1.4",
     "expo-linear-gradient": "^14.0.2",
-    "expo-status-bar": "~2.0.1",
+    "expo-status-bar": "~2.2.3",
     "firebase": "^11.4.0",
     "lucide-react-native": "^0.507.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-native": "0.76.7",
-    "react-native-gesture-handler": "^2.24.0",
+    "react": "19.0.0",
+    "react-dom": "19.0.0",
+    "react-native": "0.79.2",
+    "react-native-gesture-handler": "~2.24.0",
     "react-native-modal-datetime-picker": "^18.0.0",
     "react-native-reanimated": "^3.17.1",
     "react-native-safe-area-context": "^5.3.0",
     "react-native-screens": "^4.9.2",
-    "react-native-svg": "^15.11.2",
+    "react-native-svg": "15.11.2",
     "react-native-vector-icons": "^10.2.0",
-    "react-native-web": "~0.19.13"
+    "react-native-web": "^0.20.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
-    "@types/react": "~18.3.12",
+    "@expo/cli": "^0.24.13",
+    "@types/react": "~19.0.10",
+    "react-native-svg-transformer": "^1.5.1",
     "typescript": "^5.3.3"
   },
   "private": true
