+++ src/components/SearchBar/SearchBar.tsx
@@ -1,0 +1,17 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+import Logo from '../../../assets/logo.svg'
+
+const SearchBar = () => {
+    return (
+        <View>
+            <View>
+                <Logo width={100} height={100} />
+            </View>
+            <View></View>
+            <View></View>
+        </View>
+    )
+}
+
+export default SearchBar
