+++ types.d.ts
@@ -1,0 +1,6 @@
+declare module '*.svg' {
+    import * as React from 'react';
+    import { SvgProps } from 'react-native-svg';
+    const content: React.FC<SvgProps>;
+    export default content;
+}
